+++ src/util/Config.cpp
@@ -1,44 +1,44 @@
 #include "Config.h"
 
-void ExtractColorFromString(std::string in, CRGBA &out)
-{
-    // Remove whitespace
-    in.erase(std::remove_if(in.begin(), in.end(), isspace), in.end());
-
-    bool didR = false, didG = false, didB = false, didA = false;
-
-    size_t pos = 0;
-    for (unsigned char i = 0; i < 4; i++)
-    {
-        pos = in.find(",");
-
-        if (!didR)
-        {
-            out.r = (unsigned char)std::stoi(in.substr(0, pos).c_str());
-            didR = true;
-        }
-        else if (!didG)
-        {
-            out.g = (unsigned char)std::stoi(in.substr(0, pos).c_str());
-            didG = true;
-        }
-        else if (!didB)
-        {
-            out.b = (unsigned char)std::stoi(in.substr(0, pos).c_str());
-            didB = true;
-        }
-        else if (!didA)
-        {
-            out.a = (unsigned char)std::stoi(in.substr(0, pos).c_str());
-            didA = true;
-        }
-
-        in.erase(0, pos + 1);
-    }
-}
-
 namespace util
 {
+    void ExtractColorFromString(std::string in, CRGBA &out)
+    {
+        // Remove whitespace
+        in.erase(std::remove_if(in.begin(), in.end(), isspace), in.end());
+
+        bool didR = false, didG = false, didB = false, didA = false;
+
+        size_t pos = 0;
+        for (unsigned char i = 0; i < 4; i++)
+        {
+            pos = in.find(",");
+
+            if (!didR)
+            {
+                out.r = (unsigned char)std::stoi(in.substr(0, pos).c_str());
+                didR = true;
+            }
+            else if (!didG)
+            {
+                out.g = (unsigned char)std::stoi(in.substr(0, pos).c_str());
+                didG = true;
+            }
+            else if (!didB)
+            {
+                out.b = (unsigned char)std::stoi(in.substr(0, pos).c_str());
+                didB = true;
+            }
+            else if (!didA)
+            {
+                out.a = (unsigned char)std::stoi(in.substr(0, pos).c_str());
+                didA = true;
+            }
+
+            in.erase(0, pos + 1);
+        }
+    }
+
     Config::Config(const char *filename)
     {
         mINI::INIFile file(filename);
